rename gen_emoji into make_emoji
