+++ src/commands/ai/make_emoji.ts
